add Jenkins build configuration file
The Jenkins GitHub Pull Request Builder plugin configures a "bot"
behavior that makes a comment in each pull request to get approval to
build the branch and run the tests. Because out application is open
source we need this to avoid malicious builds. If the application were
closed source we could simply have Jenkins build every branch.

The Jenkins config requires a GitHub API key. See the line marked
'REDACTED' for where it goes in the configuration.